- added auth method - added jest tests - bug fixes - other features

+++ src/controllers/cards.ts
@@ -1,35 +1,100 @@
-import { NextFunction, Response } from 'express';
+import { Error as MongooseError } from 'mongoose';
+import { Response, NextFunction } from 'express';
 import cardsModel from '../models/card';
 import { IJwtUserSignature, RequestOrRequestWithJwt } from '../types/user';
 import { isValidJwsUserSignature } from '../utils/validation/user';
+import NotAuthorizedError from '../errors/not-authorized-error';
+import handleMongooseError from '../utils/handlers/mongoose-error-handler';
+import NotFoundError from '../errors/not-found-error';
+import {
+  AUTHORIZATION_ERROR_MESSAGE,
+  CARD_FOR_DELETION_NOT_FOUND_ERROR_MESSAGE,
+  INSUFFICIENT_PERMISSIONS_ERROR_MESSAGE,
+  UNAUTHORIZED_CARD_CREATION_ERROR_MESSAGE,
+  UNAUTHORIZED_CARD_DELETION_ERROR_MESSAGE,
+} from '../constants/error';
+import { CARD_SUCCESS_DELETION_MESSAGE } from '../constants/card';
 
+/**
+ * Получает список всех карточек из базы данных.
+ *
+ * Этот асинхронный обработчик маршрута используется для получения и отправки
+ * списка всех карточек, хранящихся в базе данных. Он не принимает никаких
+ * параметров в запросе и возвращает массив карточек.
+ *
+ * @param {_} _ - Объект запроса Express. Не используется в данной функции.
+ * @param {Response} res - Объект ответа Express, используется для отправки
+ *   ответа.
+ * @param {NextFunction} next - Функция Express `next`, используется для
+ *   передачи ошибки следующему обработчику.
+ *
+ * @returns {Promise<void>} Promise, который разрешается после отправки ответа
+ *   или передачи ошибки.
+ *
+ * @example
+ * // Пример маршрута, использующего этот обработчик:
+ * router.get('/cards', getCards);
+ */
 export const getCards = async (
-  req: RequestOrRequestWithJwt,
+  _: RequestOrRequestWithJwt,
   res: Response,
   next: NextFunction,
-) => {
+): Promise<void> => {
   try {
     const cards = await cardsModel.find({}, {});
-    res.send({ data: cards });
+    res.send({ cards });
   } catch (error) {
+    if (error instanceof MongooseError) {
+      handleMongooseError(error, next);
+      return;
+    }
     next(error);
   }
 };
 
+/**
+ * Создает новую карточку в базе данных на основе данных запроса.
+ *
+ * Этот асинхронный обработчик маршрута используется для создания новой
+ * карточки. Он требует наличия в запросе тела (body) с данными для карточки и
+ * аутентификации пользователя. При успешном создании возвращает созданную
+ * карточку.
+ *
+ * @param {RequestOrRequestWithJwt} req - Объект запроса Express, содержащий
+ *   тело запроса и данные пользователя.
+ * @param {Response} res - Объект ответа Express, используется для отправки
+ *   ответа.
+ * @param {NextFunction} next - Функция Express `next`, используется для
+ *   передачи ошибки следующему обработчику.
+ *
+ * @throws {NotAuthorizedError} Если пользователь не авторизован.
+ * @throws {MongooseError} Ошибки, связанные с операциями MongoDB.
+ *
+ * @returns {Promise<void>} Promise, который разрешается после отправки ответа
+ *   или передачи ошибки.
+ *
+ * @example
+ * // Пример маршрута, использующего этот обработчик:
+ * router.post('/cards', createCard);
+ */
 export const createCard = async (
   req: RequestOrRequestWithJwt,
   res: Response,
   next: NextFunction,
-) => {
+): Promise<void> => {
   try {
-    if (!isValidJwsUserSignature(req)) {
-      return next(new Error('Необходима авторизация'));
+    if (!isValidJwsUserSignature(req.user)) {
+      next(new NotAuthorizedError(UNAUTHORIZED_CARD_CREATION_ERROR_MESSAGE));
+      return;
     }
-    const { _id } = req.user as IJwtUserSignature;
-    const createdCard = await cardsModel.create({ ...req.body, owner: _id });
-    return res.send(createdCard);
+    const createdCard = await cardsModel.create({ ...req.body, owner: req.user._id });
+    res.send(createdCard);
   } catch (error) {
-    return next(error);
+    if (error instanceof MongooseError) {
+      handleMongooseError(error, next);
+      return;
+    }
+    next(error);
   }
 };
 
@@ -37,16 +102,33 @@ export const deleteCard = async (
   req: RequestOrRequestWithJwt,
   res: Response,
   next: NextFunction,
-) => {
+): Promise<void> => {
   try {
-    const cardId = req.params?.cardId;
-    const deleteResults = await cardsModel.deleteOne({ _id: cardId });
-    if (deleteResults.deletedCount < 1) {
-      res.status(404).send('Карточка с указанным _id не найдена.');
+    if (!isValidJwsUserSignature(req.user)) {
+      next(new NotAuthorizedError(UNAUTHORIZED_CARD_DELETION_ERROR_MESSAGE));
       return;
     }
-    res.send('Карточка была удалена');
+    const cardId = req.params?.cardId;
+    const cardToDelete = await cardsModel.findById(cardId);
+    if (!cardToDelete) {
+      next(new NotFoundError(CARD_FOR_DELETION_NOT_FOUND_ERROR_MESSAGE));
+      return;
+    }
+    if (cardToDelete.owner !== req.user._id) {
+      next(new NotAuthorizedError(INSUFFICIENT_PERMISSIONS_ERROR_MESSAGE));
+      return;
+    }
+    const deleteResults = await cardsModel.deleteOne({ _id: cardId });
+    if (deleteResults.deletedCount < 1) {
+      next(new NotFoundError(CARD_FOR_DELETION_NOT_FOUND_ERROR_MESSAGE));
+      return;
+    }
+    res.send(CARD_SUCCESS_DELETION_MESSAGE);
   } catch (error) {
+    if (error instanceof MongooseError) {
+      handleMongooseError(error, next);
+      return;
+    }
     next(error);
   }
 };
@@ -55,10 +137,11 @@ export const addLikeToCard = async (
   req: RequestOrRequestWithJwt,
   res: Response,
   next: NextFunction,
-) => {
+): Promise<void> => {
   try {
-    if (!isValidJwsUserSignature(req)) {
-      return next(new Error('Пользователь не авторизован'));
+    if (!isValidJwsUserSignature(req.user)) {
+      next(new Error(AUTHORIZATION_ERROR_MESSAGE));
+      return;
     }
     const { _id } = req.user as IJwtUserSignature;
     const updatedCard = await cardsModel.findByIdAndUpdate(
@@ -66,28 +149,41 @@ export const addLikeToCard = async (
       { $addToSet: { likes: _id } },
       { new: true },
     );
-    return res.send(updatedCard);
+    res.send(updatedCard);
   } catch (error) {
-    return next(error);
+    if (error instanceof MongooseError) {
+      handleMongooseError(error, next);
+      return;
+    }
+    next(error);
   }
 };
 export const deleteLikeFromCard = async (
   req: RequestOrRequestWithJwt,
   res: Response,
   next: NextFunction,
-) => {
+): Promise<void> => {
   try {
-    if (!isValidJwsUserSignature(req)) {
-      return next(new Error('Пользователь не авторизован'));
+    if (!isValidJwsUserSignature(req.user)) {
+      next(new Error(AUTHORIZATION_ERROR_MESSAGE));
+      return;
     }
     const { _id } = req.user as IJwtUserSignature;
     const { cardId } = req.params;
-    await cardsModel.findByIdAndUpdate(cardId, {
-      $pull: { likes: _id },
-    });
+    await cardsModel.findByIdAndUpdate(
+      cardId,
+      {
+        $pull: { likes: _id },
+      },
+      { new: true },
+    );
     const updatedCard = await cardsModel.findById(cardId);
-    return res.send(updatedCard);
+    res.send(updatedCard);
   } catch (error) {
-    return next(error);
+    if (error instanceof MongooseError) {
+      handleMongooseError(error, next);
+      return;
+    }
+    next(error);
   }
 };
